split out TestHelpers into separate parts
