+++ test/testhelpers/OffsetArrays.jl
@@ -1,52 +1,11 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-module TestHelpers
-
-using Serialization
-
-include("dimensionful.jl")
-export Furlong
-
-function open_fake_pty()
-    @static if Sys.iswindows()
-        error("Unable to create a fake PTY in Windows")
-    end
-
-    O_RDWR = Base.Filesystem.JL_O_RDWR
-    O_NOCTTY = Base.Filesystem.JL_O_NOCTTY
-
-    fdm = ccall(:posix_openpt, Cint, (Cint,), O_RDWR|O_NOCTTY)
-    fdm == -1 && error("Failed to open PTY master")
-    rc = ccall(:grantpt, Cint, (Cint,), fdm)
-    rc != 0 && error("grantpt failed")
-    rc = ccall(:unlockpt, Cint, (Cint,), fdm)
-    rc != 0 && error("unlockpt")
-
-    fds = ccall(:open, Cint, (Ptr{UInt8}, Cint),
-        ccall(:ptsname, Ptr{UInt8}, (Cint,), fdm), O_RDWR|O_NOCTTY)
-
-    # slave
-    slave = RawFD(fds)
-    master = Base.TTY(RawFD(fdm); readable = true)
-    slave, master
-end
-
-function with_fake_pty(f)
-    slave, master = open_fake_pty()
-    try
-        f(slave, master)
-    finally
-        ccall(:close,Cint,(Cint,),slave) # XXX: this causes the kernel to throw away all unread data on the pty
-        close(master)
-    end
-end
-
 # OffsetArrays (arrays with indexing that doesn't start at 1)
 
 # This test file is designed to exercise support for generic indexing,
 # even though offset arrays aren't implemented in Base.
 
-module OAs
+module OffsetArrays
 
 using Base: Indices, IndexCartesian, IndexLinear, tail
 
@@ -174,26 +133,5 @@ indslength(i::Integer) = i
 
 
 Base.resize!(A::OffsetVector, nl::Integer) = (resize!(A.parent, nl); A)
-
-end
-
-# Mimic a quantity with a physical unit that is not convertible to a real number
-struct PhysQuantity{n,T}   # n is like the exponent of the unit
-    val::T
-end
-PhysQuantity{n}(x::T) where {n,T} = PhysQuantity{n,T}(x)
-Base.zero(::Type{PhysQuantity{n,T}}) where {n,T} = PhysQuantity{n,T}(zero(T))
-Base.zero(x::PhysQuantity) = zero(typeof(x))
-Base.:+(x::PhysQuantity{n}, y::PhysQuantity{n}) where n = PhysQuantity{n}(x.val + y.val)
-Base.:-(x::PhysQuantity{n}, y::PhysQuantity{n}) where n = PhysQuantity{n}(x.val - y.val)
-Base.:*(x::PhysQuantity{n,T}, y::Int) where {n,T} = PhysQuantity{n}(x.val*y)
-Base.:/(x::PhysQuantity{n,T}, y::Int) where {n,T} = PhysQuantity{n}(x.val/y)
-Base.:*(x::PhysQuantity{n1,S}, y::PhysQuantity{n2,T}) where {n1,n2,S,T} =
-    PhysQuantity{n1+n2}(x.val*y.val)
-Base.:/(x::PhysQuantity{n1,S}, y::PhysQuantity{n2,T}) where {n1,n2,S,T} =
-    PhysQuantity{n1-n2}(x.val/y.val)
-Base.convert(::Type{PhysQuantity{0,T}}, x::Int) where T = PhysQuantity{0}(convert(T, x))
-Base.convert(::Type{P}, ::Int) where P<:PhysQuantity =
-    error("Int is incommensurate with PhysQuantity")
 
 end
